requirements.txt
asgiref==3.2.10
awsebcli==3.19.2
boto==2.49.0
boto3==1.16.30
botocore==1.19.30
cement==2.8.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
django-storages==1.10.1
django-widget-tweaks==1.4.8
future==0.16.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pypiwin32==223
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.1
pywin32==300
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
whitenoise==5.2.0
